# Avoid type casting